pyproject.toml
[project]
name = "paradex-lighter-arb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "lighter-sdk",
    "requests>=2.32.5",
]

[tool.uv.sources]
lighter-sdk = { git = "https://github.com/elliottech/lighter-python.git" }
